LPD-66497 Move the refactored code to a better place
