+++ modules/test/playwright/tests/frontend-editor-ckeditor-web/main/tests/ckeditor4/legacy.spec.ts
@@ -7,8 +7,8 @@ import {expect, mergeTests} from '@playwright/test';
 
 import {featureFlagsTest} from '../../../../../fixtures/featureFlagsTest';
 import {loginTest} from '../../../../../fixtures/loginTest';
-import {ckeditorSamplePageTest} from '../../fixtures/ckeditorSamplePageTest';
-import {SubTabName, TabName} from "../../pages/CKEditorSamplePage";
+import {ckeditorSamplePageTest} from '../../../../frontend-editor-ckeditor-sample-web/fixtures/ckeditorSamplePageTest';
+import {SubTabName, TabName} from "../../../../frontend-editor-ckeditor-sample-web/pages/CKEditorSamplePage";
 
 export const test = mergeTests(
 	ckeditorSamplePageTest,
